feat: criando hook useSummary para encapsular a logica de calculo do summary
