+++ ignite-reactjs-2023/04-dt-money/src/components/Summary/index.tsx
@@ -1,33 +1,10 @@
 import { ArrowCircleDown, ArrowCircleUp, CurrencyDollar } from "phosphor-react";
 import { SummaryCard, SummaryContainer } from "./styles";
-import { TransactionsContext } from "../../contexts/TransactionsContext";
-import { useContext } from "react";
 import { priceFormatter } from "../../utils/formatter";
+import { useSummary } from "../../hooks/useSummary";
 
 export function Summary(){
-    const {transactions} = useContext(TransactionsContext);
-
-    const summary = transactions.reduce(
-        (acc, transaction) => {
-            switch (transaction.type){
-                case 'income':
-                    acc.income += transaction.price;
-                    acc.total += transaction.price;
-                    break;
-                case 'outcome':
-                    acc.outcome += transaction.price;
-                    acc.total -= transaction.price;
-                    break;
-                default:
-                    break;
-            }
-            return acc;
-        },{
-            income: 0,
-            outcome: 0,
-            total: 0
-        }
-    );
+    const summary = useSummary();
 
     return (
         <SummaryContainer>
